requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
phonenumbers==9.0.0
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.3
